Encapsulate FlushCompletionTracker semaphore; move to own file
Replace SemaphoreSlim parameter with bool enableThrottling so the
semaphore is a private implementation detail. Move FlushCompletionTracker
from PageAsyncResultTypes.cs to its own file.
