+++ libs/storage/Tsavorite/cs/src/core/Utilities/PageAsyncResultTypes.cs
@@ -7,7 +7,6 @@ using System;
 using System.Diagnostics;
 using System.Runtime.InteropServices;
 using System.Threading;
-using System.Threading.Tasks;
 
 namespace Tsavorite.core
 {
@@ -73,61 +72,6 @@ namespace Tsavorite.core
     /// <summary>
     /// Shared flush completion tracker, when bulk-flushing many pages
     /// </summary>
-    internal sealed class FlushCompletionTracker
-    {
-        /// <summary>
-        /// Task completion source to signal when all page flushes are done, or to fault on error.
-        /// </summary>
-        readonly TaskCompletionSource<bool> completionTcs;
-
-        /// <summary>
-        /// Semaphore to wait on for per-page flush completion (throttling)
-        /// </summary>
-        readonly SemaphoreSlim flushSemaphore;
-
-        /// <summary>
-        /// Number of pages being flushed
-        /// </summary>
-        int count;
-
-        public override string ToString()
-        {
-            var flushSemCount = flushSemaphore?.CurrentCount.ToString() ?? "null";
-            return $"count {count}, flushSemCount {flushSemCount}";
-        }
-
-        /// <summary>
-        /// Create a flush completion tracker
-        /// </summary>
-        /// <param name="completionTcs">TaskCompletionSource to signal when all flushes complete or to fault on error</param>
-        /// <param name="flushSemaphore">Semaphore to release when each flush completes (for throttling)</param>
-        /// <param name="count">Number of pages to flush</param>
-        public FlushCompletionTracker(TaskCompletionSource<bool> completionTcs, SemaphoreSlim flushSemaphore, int count)
-        {
-            this.completionTcs = completionTcs;
-            this.flushSemaphore = flushSemaphore;
-            this.count = count;
-        }
-
-        /// <summary>
-        /// Complete flush of one page
-        /// </summary>
-        public void CompleteFlush()
-        {
-            _ = (flushSemaphore?.Release());
-            if (Interlocked.Decrement(ref count) == 0)
-                _ = completionTcs.TrySetResult(true);
-        }
-
-        /// <summary>
-        /// Signal that the flush failed with an exception.
-        /// </summary>
-        public void SetException(Exception ex)
-            => _ = completionTcs.TrySetException(ex);
-
-        public void WaitOneFlush() => flushSemaphore?.Wait();
-    }
-
     internal enum FlushRequestState : byte
     {
         /// <summary>The default; we are here for <see cref="AllocatorBase{TStoreFunctions, TAllocator}.AsyncFlushPagesForReadOnly"/> flush. This
